# next/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GChat API",
        "description": "Документация API для проекта GChat",
        "version": "1.0.0"
    },
    "paths": {
...
        "/tags": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Список тегов",
                "operationId": "listTags",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/GetTags_CursorParam"
                    },
                    {
                        "$ref": "#/components/parameters/GetTags_LimitParam"
                    },
                    {
                        "$ref": "#/components/parameters/GetTags_DirectionParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/TagListCursorPaginatedDto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "Создать тег",
                "operationId": "createTag",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/TagDto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    }
                }
            }
        },
        "/tags/{id}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Получить тег",
                "operationId": "getTag",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TagIdParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/TagDto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tag"
                ],
                "summary": "Обновить тег",
                "operationId": "updateTag",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TagIdParam"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/TagDto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tag"
                ],
                "summary": "Удалить тег",
                "operationId": "deleteTag",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TagIdParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            }
        },
        "/api/translate": {
            "get": {
                "tags": [
                    "Translator"
                ],
                "summary": "Перевести текст",
                "operationId": "translate",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TranslateRequest_LangParam"
                    },
                    {
                        "$ref": "#/components/parameters/TranslateRequest_TextParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "translation": {
                                                            "type": "string",
                                                            "example": "Translated text"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerErrorResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TagDto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "b8f3a1c2-1234-4a5b-9cde-112233445566"
                    },
                    "name": {
                        "type": "string",
                        "example": "vip"
                    },
                    "color": {
                        "type": "string",
                        "example": "#FF9900",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CursorPaginatedDto": {
                "description": "Generic cursor-based paginated response wrapper",
                "properties": {
                    "items": {
                        "description": "List of items for the current page",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "nextCursor": {
                        "type": "string",
                        "example": "eyJpZCI6Ij...",
                        "nullable": true
                    },
                    "prevCursor": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "TagListCursorPaginatedDto": {
                "description": "Cursor-based paginated response with TagDto items",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/CursorPaginatedDto"
                    },
                    {
                        "properties": {
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/TagDto"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "StoreTagRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "color": {
                        "type": "string",
                        "maxLength": 32,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateTagRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "color": {
                        "type": "string",
                        "maxLength": 32,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Request successful"
                    },
                    "data": {
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Meta"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Meta": {
                "properties": {
                    "next_cursor": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "prev_cursor": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "BadRequestResponse": {
                "description": "Bad request",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/ApiResponse"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid request parameters"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "field": "Validation error message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    }
                }
            },
...
            "NotFoundResponse": {
                "description": "Resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/ApiResponse"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    }
                }
            },
            "ServerErrorResponse": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/ApiResponse"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    }
                }
            },
...
            "ValidationErrorResponse": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/ApiResponse"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "field1": [
                                                    "The field1 is required."
                                                ],
                                                "field2": [
                                                    "The field2 must be an integer."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    }
                }
            }
        },
        "parameters": {
...
            "GetTags_CursorParam": {
                "name": "cursor",
                "in": "query",
                "description": "Cursor for pagination",
                "schema": {
                    "type": "string",
                    "nullable": true
                }
            },
            "GetTags_LimitParam": {
                "name": "limit",
                "in": "query",
                "description": "Items per page (1..200)",
                "schema": {
                    "type": "integer",
                    "default": 50
                }
            },
            "GetTags_DirectionParam": {
                "name": "direction",
                "in": "query",
                "description": "Pagination direction",
                "schema": {
                    "type": "string",
                    "default": "down",
                    "enum": [
                        "up",
                        "down"
                    ]
                }
            },
            "TagIdParam": {
                "name": "id",
                "in": "path",
                "description": "Tag UUID",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "uuid"
                }
            },
            "TranslateRequest_LangParam": {
                "name": "lang",
                "in": "query",
                "description": "Язык перевода",
                "required": true,
                "schema": {
                    "type": "string",
                    "maxLength": 32
                }
            },
            "TranslateRequest_TextParam": {
                "name": "text",
                "in": "query",
                "description": "Текст для перевода",
                "required": true,
                "schema": {
                    "type": "string",
                    "maxLength": 8192
                }
            }
        }
    },
...
}